refactor(api): split security.ts into focused modules (#1611)
Splits api/src/security.ts (1,339 LOC) into focused modules under
api/src/security/ to shrink review-blast-radius for upcoming auth work.

Module layout:
- types.ts         - Shared types + STATE_CHANGING_METHODS
- cookies.ts       - Cookie extraction + Set-Cookie header builders
- sessions.ts      - In-memory session invalidation tracking
- tokens.ts        - HMAC sign/verify + session resolver + refresh
- csrf.ts          - Origin-based CSRF middleware + helpers
- middleware.ts    - Auth gate + write auth middleware
- config.ts        - Environment-driven security config resolution
- index.ts         - Barrel re-exports preserving the public surface

No behavior changes. All existing imports of ./security keep working via
the barrel. Every file is <400 LOC. The TODO at the in-memory
invalidation scan is preserved in sessions.ts and remains tracked by
#1269.
